finish changing from variant to union
