+++ pdqcif/pdqcif.vcxproj
@@ -147,7 +147,6 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="src\cif.cpp" />
     <ClCompile Include="src\cif_convert.cpp" />
     <ClCompile Include="src\test_main.cpp" />
     <ClCompile Include="src\vendor\glm\detail\glm.cpp" />
@@ -165,7 +164,7 @@
   <ItemGroup>
     <ClInclude Include="src\cif.h" />
     <ClInclude Include="src\cif_convert.h" />
-    <ClInclude Include="src\cif_rules.h" />
+    <ClInclude Include="src\cif_file.h" />
     <ClInclude Include="src\vendor\glm\common.hpp" />
     <ClInclude Include="src\vendor\glm\detail\compute_common.hpp" />
     <ClInclude Include="src\vendor\glm\detail\compute_vector_relational.hpp" />
